Introduce a mdextract command that writes snippet files to a target dir.
