+++ package.json
@@ -4,13 +4,9 @@
   "description": "The MarkDown snippet injector generates MD code snippets by extracting them from the source code of your projects.",
   "main": "index.js",
   "bin": {
-    "mdinject": "index.js"
+    "mdinject": "index.js",
+    "mdextract": "extract.js"
   },
-  "files": [
-    "index.ts",
-    "index.d.ts",
-    "tsconfig.json"
-  ],
   "typescript": {
     "definition": "index.d.ts"
   },
